+++ codegen/src/contributing/mod.rs
@@ -1,6 +1,9 @@
-use crate::markdown::{table_of_contents, update_markdown_section};
+use crate::{
+    error::CodegenError,
+    markdown::{table_of_contents, update_markdown_section},
+};
 
-pub fn generate() -> anyhow::Result<()> {
+pub fn generate() -> Result<(), CodegenError> {
     let path = std::path::PathBuf::from("./CONTRIBUTING.md");
 
     let mut contents = std::fs::read_to_string(&path)?;